requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
keras==2.15.0
kiwisolver==1.4.5
langdetect==1.0.9
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
mplcursors==0.5.3
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
pkgconfig==1.5.5
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydyf==0.8.0
pyparsing==3.1.1
pypdfium2==4.27.0
pyphen==0.14.0
python-dateutil==2.8.2
python-doctr==0.7.0
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf2onnx==1.16.1
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==1.26.18
weasyprint==61.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
zopfli==0.2.3
